Rename benchmark timer/particle generation
