+++ core/performance_test/benchmark/Cabana_CommPerformance.cpp
@@ -9,7 +9,7 @@
  * SPDX-License-Identifier: BSD-3-Clause                                    *
  ****************************************************************************/
 
-#include "Cabana_BenchmarkTimer.hpp"
+#include "Cabana_BenchmarkUtils.hpp"
 
 #include <Cabana_Core.hpp>
